Allow specifying the connection to use for SMTP
This is a simple change which shouldn't change any existing implementation (backwards compatible) but it should allow you to use a new function to specify your own TCP dialer for the SMTP client. This allows us to start using a SOCKS5 proxy as discussed before for making SMTP connections from Desk so we can use one IP to send all emails and tell our customers to whitelist this IP address via CNAME or otherwise.